requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.0
pydantic==2.11.4
supabase==2.15.1
python-multipart==0.0.20
httpx==0.27.0
aiofiles==23.2.1
pillow==10.1.0
python-dateutil==2.8.2
pydantic-settings==2.0.3
openai>=1.0.0
numpy>=1.26.0